+++ app/src/main/java/com/example/planlekcji/ckziu_elektryk/client/CKZiUElektrykClient.java
@@ -4,10 +4,10 @@ import com.example.planlekcji.ckziu_elektryk.client.article.ArticleService;
 import com.example.planlekcji.ckziu_elektryk.client.article.ArticleServiceFactory;
 import com.example.planlekcji.ckziu_elektryk.client.calendar.CalenderService;
 import com.example.planlekcji.ckziu_elektryk.client.calendar.CalenderServiceFactory;
-import com.example.planlekcji.ckziu_elektryk.client.replacments.ReplacementService;
+import com.example.planlekcji.ckziu_elektryk.client.replacements.ReplacementService;
 import com.example.planlekcji.ckziu_elektryk.client.timetable.SchoolEntryType;
 import com.example.planlekcji.ckziu_elektryk.client.timetable.TimetableService;
-import com.example.planlekcji.ckziu_elektryk.client.replacments.ReplacementServiceFactory;
+import com.example.planlekcji.ckziu_elektryk.client.replacements.ReplacementServiceFactory;
 import com.example.planlekcji.ckziu_elektryk.client.timetable.TimetableServiceFactory;
 import com.example.planlekcji.ckziu_elektryk.client.timetable.info.TimetableInfo;
 import com.example.planlekcji.ckziu_elektryk.client.timetable.info.TimetableInfoService;
