+++ app/src/main/java/com/example/planlekcji/ckziu_elektryk/client/replacements/ReplacementChange.java
@@ -1,0 +1,4 @@
+package com.example.planlekcji.ckziu_elektryk.client.replacements;
+
+public record ReplacementChange(String period, String info) {
+}
