+++ app/src/main/java/com/example/planlekcji/ckziu_elektryk/client/replacements/ReplacementService.java
@@ -1,4 +1,4 @@
-package com.example.planlekcji.ckziu_elektryk.client.replacments;
+package com.example.planlekcji.ckziu_elektryk.client.replacements;
 
 import java.util.Date;
 import java.util.List;
