--- app/src/main/java/com/example/planlekcji/ckziu_elektryk/client/replacments/ReplacementChange.java
@@ -1,4 +1,0 @@
-package com.example.planlekcji.ckziu_elektryk.client.replacments;
-
-public record ReplacementChange(String period, String info) {
-}
